Day 11 solution
First tried with pure brute force but it obviously started failing in
second part pretty early. I thought it will be enough to just cache
results to speed up calculations but true complexity boost was coming
from array traversing and gains were minimal.

I later on figured out that my approach with inserting arrays with two
split number for case with even number of digits and then flattening the
array probably means that order of pebbles does not count in reality and
reimplemented with caching not only for value calculations but also
values with their counts in a hash map instead of storing them in long
array multiple times.

I've seen more optimised solutions in other languages running this task in
around 50ms so my 150ms-ish result in ruby is pretty satisfying to me
tbh.